requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pandas
playwright
beautifulsoup4
